moved eospac specific functionality into eospac-wrapper
