+++ singularity-eos/eos/eos_eospac.cpp
@@ -15,32 +15,9 @@
 #ifdef SINGULARITY_USE_EOSPAC
 
 #include <singularity-eos/eos/eos.hpp>
-#include <eos_Interface.h>
+#include <eospac-wrapper/eospac_wrapper.hpp>
 
-PORTABLE_INLINE_FUNCTION Real TemperatureToSesame(const Real CodeTemp) {
-  return CodeTemp;
-}
-PORTABLE_INLINE_FUNCTION Real TemperatureFromSesame(const Real SesTemp) {
-  return SesTemp;
-}
-PORTABLE_INLINE_FUNCTION Real PressureFromSesame(const Real SesPress) {
-  return 1e10 * SesPress;
-}
-PORTABLE_INLINE_FUNCTION Real PressureToSesame(const Real CodePress) {
-  return 1e-10 * CodePress;
-}
-PORTABLE_INLINE_FUNCTION Real SieToSesame(const Real CodeSie) {
-  return 1e-10 * CodeSie;
-}
-PORTABLE_INLINE_FUNCTION Real SieFromSesame(const Real SesSie) {
-  return 1e10 * SesSie;
-}
-PORTABLE_INLINE_FUNCTION Real CvFromSesame(const Real SesCv) {
-  return 1e10 * SesCv;
-}
-PORTABLE_INLINE_FUNCTION Real BulkModulusFromSesame(const Real SesBmod) {
-  return 1e10 * SesBmod;
-}
+using namespace EospacWrapper;
 
 namespace singularity {
 
@@ -48,41 +25,33 @@ namespace singularity {
 // points to the same table? Does EOSPAC give me multiple (reference-counted)
 // handles to the same table or read in the table multiple times?
 
-EOSPAC::EOSPAC(const int matid) : matid_(matid) {
-  EOS_INTEGER ntables = NT, errorCode = EOS_OK;
-  EOS_INTEGER tableType[NT] = {EOS_Pt_DT, EOS_T_DUt, EOS_Ut_DT, EOS_D_PtT,
-                               EOS_Info};
-  EOS_INTEGER MATID[NT];
-  for (int i = 0; i < NT; i++) {
-    MATID[i] = matid;
-  }
-  eos_CreateTables(&ntables, tableType, MATID, tablehandle, &errorCode);
-  // eos_SetOption(&tablehandle[0],&THE_OPTION,EOS_NullPtr,&errorCode);
-  eos_LoadTables(&ntables, tablehandle, &errorCode);
-  PofRT_table_ = tablehandle[0];
-  TofRE_table_ = tablehandle[1];
-  EofRT_table_ = tablehandle[2];
-  RofPT_table_ = tablehandle[3];
-  EOS_Info_table_ = tablehandle[4];
+EOSPAC::EOSPAC(const int matid,  bool invert_at_setup) : matid_(matid) {
+  EOS_INTEGER tableType[NT] = {EOS_Pt_DT, EOS_T_DUt, EOS_Ut_DT, EOS_D_PtT};
+  eosSafeLoad(nt, matid, tableType, tablehandle,
+	      std::vector<std::string>{"EOS_Pt_DT",
+					 "EOS_T_DUt",
+					 "EOS_Ut_DT", "EOS_D_PtT"};
+	      Verbosity::Quiet, invert_at_setup);
 
   // Set reference states
-  EOS_INTEGER error_code = EOS_OK;
-  EOS_INTEGER infoItems[1] = {EOS_Normal_Density};
-  EOS_INTEGER nitems[] = {1};
-  eos_GetTableInfo(&EOS_Info_table_, nitems, infoItems, &rho_ref_, &error_code);
+  SesameMetadata m;
+  eosGetMetadata(matid, metadata, verbosity::Quiet);
+  rho_ref_ = m.normalDensity;
 
   EOS_REAL R[1] = {rho_ref_};
   EOS_REAL T[1] = {TemperatureToSesame(temp_ref_)};
   EOS_REAL E[1], P[1], dx[1], dy[1];
   EOS_REAL DEDR, DEDT, DPDR, DPDT, DPDE, BMOD;
   EOS_INTEGER nxypairs = 1;
-  eos_Interpolate(&EofRT_table_, &nxypairs, R, T, E, dx, dy, &errorCode);
+  eosSafeInterpolate(&EofRT_table_, nxypairs, R, T, E, dx, dy, "EofRT",
+		     Verbosity::Quiet);
   DEDR = dx[0];
   DEDT = dy[0];
   sie_ref_ = SieFromSesame(E[0]);
   cv_ref_ = CvFromSesame(std::max(DEDT, 0.0));
 
-  eos_Interpolate(&PofRT_table_, &nxypairs, R, T, P, dx, dy, &errorCode);
+  eosSafeInterpolate(&PofRT_table_, nxypairs, R, T, P, dx, dy, "PofRT",
+		     Verbosity::Quiet);
   Real PRESS = P[0];
   press_ref_ = PressureFromSesame(PRESS);
   DPDR = dx[0];
@@ -98,39 +67,45 @@ EOSPAC::EOSPAC(const int matid) : matid_(matid) {
 PORTABLE_FUNCTION Real EOSPAC::TemperatureFromDensityInternalEnergy(
     const Real rho, const Real sie, Real *lambda) const {
   EOS_REAL R[1] = {rho}, E[1] = {SieToSesame(sie)}, T[1], dTdr[1], dTde[1];
-  EOS_INTEGER nxypairs = 1, errorCode;
+  EOS_INTEGER nxypairs = 1;
   EOS_INTEGER table = TofRE_table_;
-  eos_Interpolate(&table, &nxypairs, R, E, T, dTdr, dTde, &errorCode);
+  eosSafeInterpolate(&table, nxypairs, R, E, T, dTdr, dTde, "TofRE",
+		     Verbosity::Quiet);
   return Real(TemperatureFromSesame(T[0]));
 }
+
 PORTABLE_FUNCTION Real EOSPAC::PressureFromDensityTemperature(
     const Real rho, const Real temp, Real *lambda) const {
   EOS_REAL R[1] = {rho}, P[1], T[1] = {TemperatureToSesame(temp)}, dPdr[1],
            dPdT[1];
-  EOS_INTEGER nxypairs = 1, errorCode;
+  EOS_INTEGER nxypairs = 1;
   EOS_INTEGER table = PofRT_table_;
-  eos_Interpolate(&table, &nxypairs, R, T, P, dPdr, dPdT, &errorCode);
+  eosSafeInterpolate(&table, nxypairs, R, T, P, dPdr, dPdT, "PofRT",
+		     Verbosity::Quiet);
   return Real(PressureFromSesame(P[0]));
 }
+
 PORTABLE_FUNCTION void EOSPAC::FillEos(Real &rho, Real &temp, Real &sie,
                                        Real &press, Real &cv, Real &bmod,
                                        const unsigned long output,
                                        Real *lambda) const {
   EOS_REAL R[1] = {rho}, T[1] = {TemperatureToSesame(temp)}, E[1], P[1], dx[1],
            dy[1];
-  EOS_INTEGER nxypairs = 1, errorCode = EOS_OK;
+  EOS_INTEGER nxypairs = 1;
   Real CV, BMOD_T, BMOD, SIE, PRESS, DPDE, DPDT, DPDR, DEDT, DEDR;
   if ((output & thermalqs::specific_internal_energy) ||
       (output & thermalqs::specific_heat || output & thermalqs::bulk_modulus)) {
     EOS_INTEGER table = EofRT_table_;
-    eos_Interpolate(&table, &nxypairs, R, T, E, dx, dy, &errorCode);
+    eosSafeInterpolate(&table, nxypairs, R, T, E, dx, dy, "EofRT",
+		       Verbosity::Quiet);
     SIE = E[0];
     DEDR = dx[0];
     DEDT = dy[0];
   }
   if ((output & thermalqs::pressure) || (output & thermalqs::bulk_modulus)) {
     EOS_INTEGER table = PofRT_table_;
-    eos_Interpolate(&table, &nxypairs, R, T, P, dx, dy, &errorCode);
+    eosSafeInterpolate(&table, nxpairs, R, T, P, dx, dy, "PofRT",
+		       Verbosity::Quiet);
     PRESS = P[0];
     DPDR = dx[0];
     DPDT = dy[0];
@@ -210,13 +185,15 @@ PORTABLE_FUNCTION Real EOSPAC::GruneisenParamFromDensityTemperature(
     const Real rho, const Real temperature, Real *lambda) const {
   EOS_REAL R[1] = {rho}, T[1] = {TemperatureToSesame(temperature)};
   EOS_REAL E[1], P[1], dx[1], dy[1];
-  EOS_INTEGER nxypairs = 1, errorCode = EOS_OK;
+  EOS_INTEGER nxypairs = 1;
   Real DPDT, DEDT, DPDE;
   EOS_INTEGER table = EofRT_table_;
-  eos_Interpolate(&table, &nxypairs, R, T, E, dx, dy, &errorCode);
+  eosSafeInterpolate(&table, nxypairs, R, T, E, dx, dy, "EofRT",
+		     Verbosity::Quiet);
   DEDT = dy[0];
   table = PofRT_table_;
-  eos_Interpolate(&table, &nxypairs, R, T, P, dx, dy, &errorCode);
+  eosSafeInterpolate(&table, nxypairs, R, T, P, dx, dy, "PofRT",
+		     Verbosity::Quiet);
   DPDT = dy[0];
   DPDE = DPDT / DEDT;
   return PressureFromSesame(SieToSesame(DPDE)) / (rho + EPS);
@@ -239,11 +216,13 @@ void EOSPAC::DensityEnergyFromPressureTemperature(const Real press, const Real t
   EOS_INTEGER table;
 
   table = RofPT_table_;
-  eos_Interpolate(&table, &nxypairs, P, T, R, dx, dy, &errorCode);
+  eosSafeInterpolate(&table, nxypairs, P, T, R, dx, dy, "RofPT",
+		     Verbosity::Quiet);
   rho = R[0];
   
   table = EofRT_table_;
-  eos_Interpolate(&table, &nxypairs, R, T, E, dx, dy, &errorCode);
+  eosSafeInterpoalte(&table, nxypairs, R, T, E, dx, dy, "EofPRT",
+		     Verbosity::Quiet);
   sie = SieFromSesame(E[0]);
 }
 
@@ -252,15 +231,17 @@ PORTABLE_FUNCTION void EOSPAC::PTofRE(const Real rho, const Real sie,
                                       Real &dpdr, Real &dpde, Real &dtdr,
                                       Real &dtde) const {
   EOS_REAL R[1] = {rho}, T[1], E[1] = {SieToSesame(sie)}, P[1], dx[1], dy[1];
-  EOS_INTEGER nxypairs = 1, errorCode = EOS_OK;
+  EOS_INTEGER nxypairs = 1;
   Real DTDR_E, DTDE_R, DPDR_T, DPDT_R;
 
   EOS_INTEGER table = TofRE_table_;
-  eos_Interpolate(&table, &nxypairs, R, E, T, dx, dy, &errorCode);
+  eosSafeInterpolate(&table, nxypairs, R, E, T, dx, dy, "TofRE",
+		     Verbosity::Quiet);
   DTDR_E = dx[0];
   DTDE_R = dy[0];
   table = PofRT_table_;
-  eos_Interpolate(&table, &nxypairs, R, T, P, dx, dy, &errorCode);
+  eosSafeInterpolate(&table, nxypairs, R, T, P, dx, dy, "PofRT",
+		     Verbosity::Quiet);
   DPDR_T = dx[0];
   DPDT_R = dy[0];
